30 mayıs

+++ Entities/Concrete/OrderDetail.cs
@@ -8,6 +8,6 @@ namespace Entities.Concrete
         public int OrderId { get; set; }
         public int ProductId { get; set; }
         public int Count { get; set; }
-        public int SalePrice { get; set; }
+        public decimal SalePrice { get; set; }
     }
 }
